test-minions-observe-llm/proxyApp/app/requirements.txt
fastapi
uvicorn
httpx
pydantic
prometheus_client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-logging
openlit
pyyaml
